[FEAT] 자유교환 받은 요청 조회
