--- src/main/java/com/app/treen/chat_room/dto/response/ChatRoomCreateResponse.java
@@ -1,13 +1,0 @@
-package com.app.treen.chat_room.dto.response;
-
-import lombok.*;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class ChatRoomCreateResponse {
-
-    private long chatRoomId;
-}
