+++ src/main/java/com/app/treen/jpa/repository/trade/TradeChatRoomRepository.java
@@ -1,0 +1,9 @@
+package com.app.treen.jpa.repository.trade;
+
+import com.app.treen.trade.trade_chat.chat_room.TradeChatRoom;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TradeChatRoomRepository extends JpaRepository<TradeChatRoom, Long> {
+
+
+}
