+++ src/main/java/com/app/treen/jpa/repository/transactions/TransactionsRepository.java
@@ -1,6 +1,6 @@
 package com.app.treen.jpa.repository.transactions;
 
-import com.app.treen.chat_room.entity.ChatRoom;
+import com.app.treen.transactions.transactions_chat.chat_room.entity.ChatRoom;
 import com.app.treen.transactions.entity.Transactions;
 import org.springframework.data.jpa.repository.JpaRepository;
