+++ src/main/java/com/app/treen/transactions/transactions_chat/chat_room/dto/response/ChatRoomCreateResponse.java
@@ -1,0 +1,13 @@
+package com.app.treen.transactions.transactions_chat.chat_room.dto.response;
+
+import lombok.*;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class ChatRoomCreateResponse {
+
+    private long chatRoomId;
+}
