+++ src/main/java/com/app/treen/transactions/transactions_chat/message/dto/MessageResponse.java
@@ -1,0 +1,17 @@
+package com.app.treen.transactions.transactions_chat.message.dto;
+
+import lombok.*;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class MessageResponse {
+
+    private Long roomId;
+
+    private String type;
+
+    private String content;
+}
